backend/requirements.txt
fastapi[standard]>=0.111.0
uvicorn>=0.30.1
pydantic>=2.7.4
python-dotenv>=1.0.1
langchain>=0.2.6
langchain-community>=0.2.6
langchain-core>=0.2.10
langgraph>=0.1.4
langchain-openai>=0.1.13
langchain-google-genai>=1.0.7
langchain-anthropic>=0.1.19
httpx>=0.27.0
supabase>=2.5.1
